% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area2d_kde_on_point.R
\name{area2d_kde_on_point}
\alias{area2d_kde_on_point}
\title{Title}
\usage{
area2d_kde_on_point(
  Points = as.matrix(sf::st_coordinates(Hausnummern)),
  Pointnames = dplyr::pull(Hausnummern, Adresse),
  polygons = STT,
  col = "Straßenkriminalität",
  burnin = 5,
  samples = 10,
  gridsize = 100
)
}
\arguments{
\item{gridsize}{}
}
\description{
Title
}
