% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_expr.R
\name{left_join_expr}
\alias{left_join_expr}
\title{Expression Left-Join}
\usage{
left_join_expr(
  data,
  data2,
  key_name = "by_key",
  data_expr,
  data2_expr,
  use_cols = NULL
)
}
\arguments{
\item{data}{A data frame.}

\item{data2}{A second data frame.}

\item{key_name}{A name for a interim key variable which is automatically removed afterwards. Is named "by_key" by default and should be changed if this column name is already present in data or data2.}

\item{data_expr}{A expression in a string containing the directions to produce the key variable in data.}

\item{data2_expr}{A expression in a string containing the directions to produce the key variable in data2.}

\item{use_cols}{A optional character vector. Specify the columns to extract in data2.}
}
\description{
Left-Join a data frame with another based on parsed expressions.
}
\details{
This function has been designed as a helper function to join a data frame with another based on
key column which is by expressions contained in a string defined for each data frame.
This may help to avoid to bloat the code with multiple data-pre-processing-steps by altering the structure
of the data frames to make them compatible for dplyrs left_join-function.
}
